Re-enable the CameraServer, deprecate VisionAuto
After some chatting we have decided to move the camera from the pi to the rio in an attempt to enhance performance and quality on the driver station. This sacrifices vision processing (auto and SmartDash feedback) for now, but we may be able to serve the camera stream from the rio as an endpoint that the pi can connect to and process.